Move code into a proper package
