+++ polypheny/__init__.py
@@ -1,0 +1,54 @@
+from polypheny.connection import Connection, Cursor
+from polypheny.exceptions import *
+
+import datetime
+
+apilevel = '2.0'
+threadsafety = 0
+paramstyle = 'qmark'
+
+
+def Date(year, month, day):
+    return datetime.date(year, month, day)
+
+
+def Time(hour, minute, second):
+    return datetime.time(hour, minute, second)
+
+
+def Timestamp(year, month, day, hour, minute, second):
+    return datetime.datetime(year, month, day, hour, minute, second)
+
+
+# See PEP 249
+import time
+
+
+def DateFromTicks(ticks):
+    return Date(*time.localtime(ticks)[:3])  # TODO: Really local time?
+
+
+def TimeFromTicks(ticks):
+    return Time(*time.localtime(ticks)[3:6])
+
+
+def TimestampFromTicks(ticks):
+    return Timestamp(*time.localtime(ticks)[:6])
+
+
+def Binary(string):
+    return string.encode('UTF-8')
+
+
+# Intentionally omitted, we always give type_code = None, like sqlite3
+# STRING = 1
+# BINARY = 2
+# NUMBER = 3
+# DATETIME = 4
+# ROWID = 5
+
+
+def connect(address, port, /, username, password, transport=None):
+    """ Connect to a Polypheny instance
+    """
+    return Connection(address, port, username, password, transport)
