+++ polypheny/connection.py
@@ -2,53 +2,9 @@ import datetime
 from typing import Union, List, Any, Dict
 
 import relational_frame_pb2
-import rpc
-from exceptions import *
-from serialize import *
-
-apilevel = '2.0'
-threadsafety = 0
-paramstyle = 'qmark'
-
-
-def Date(year, month, day):
-    return datetime.date(year, month, day)
-
-
-def Time(hour, minute, second):
-    return datetime.time(hour, minute, second)
-
-
-def Timestamp(year, month, day, hour, minute, second):
-    return datetime.datetime(year, month, day, hour, minute, second)
-
-
-# See PEP 249
-import time
-
-
-def DateFromTicks(ticks):
-    return Date(*time.localtime(ticks)[:3])  # TODO: Really local time?
-
-
-def TimeFromTicks(ticks):
-    return Time(*time.localtime(ticks)[3:6])
-
-
-def TimestampFromTicks(ticks):
-    return Timestamp(*time.localtime(ticks)[:6])
-
-
-def Binary(string):
-    return string.encode('UTF-8')
-
-
-# Intentionally omitted, we always give type_code = None, like sqlite3
-# STRING = 1
-# BINARY = 2
-# NUMBER = 3
-# DATETIME = 4
-# ROWID = 5
+from polypheny import rpc
+from polypheny.exceptions import *
+from polypheny.serialize import *
 
 class Connection:
     def __init__(self, address, port, username, password, transport):
@@ -326,9 +282,3 @@ class Cursor:
 
     def setoutputsize(self, sizes, column=None):
         pass  # We are free to do nothing
-
-
-def connect(address, port, /, username, password, transport=None):
-    """ Connect to a Polypheny instance
-    """
-    return Connection(address, port, username, password, transport)
